+++ app/Http/Livewire/Admin/Category/Index.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire\Admin\Category;
 use Livewire\Component;
 use App\Models\Category;
 use Livewire\WithPagination;
+use App\Models\Product;
 
 class Index extends Component
 {
@@ -21,15 +22,25 @@ class Index extends Component
     public function destroyCategory()
     {
         $category = Category::find($this->category_id);
-        $category->delete();
-        session()->flash('message','Categoria Eliminada');
-        $this->dispatchBrowserEvent('close-modal');
+        $producto = Product::all()->where('category_id','=',$this->category_id); 
+        if(count($producto)==0){ 
+            $category->delete();
+            session()->flash('message','Categoria Eliminada');
+            $this->dispatchBrowserEvent('close-modal');
+        }else{  
+            $category->status = 1;
+            $category->update();
+            session()->flash('message','Categoria Eliminada');
+            $this->dispatchBrowserEvent('close-modal');
+        }
+        
+        
     }
 
     public function render()
     {
         
-        $categories = Category::orderBy('id','DESC')->paginate(10);
+        $categories = Category::orderBy('id','DESC')->where('status','=',0)->paginate(10);
         return view('livewire.admin.category.index',['categories' => $categories]);
     }
 }
