+++ app/Http/Livewire/Admin/Company/Index.php
@@ -5,6 +5,9 @@ namespace App\Http\Livewire\Admin\Company;
 
 use Livewire\Component;
 use App\Models\Company;
+use App\Models\Ingreso;
+use App\Models\Venta;
+use App\Models\Detalleinventario;
 use Livewire\WithPagination;
 
 use Illuminate\Support\Facades\File;
@@ -25,18 +28,31 @@ class Index extends Component
     {
         $company = Company::find($this->company_id);
         $company2 =$company;
-       if( $company->delete())
+        $detalleinventario = Detalleinventario::all()->where('company_id','=',$this->company_id);  
+        $ingreso = Ingreso::all()->where('company_id','=',$this->company_id); 
+        $venta = Venta::all()->where('company_id','=',$this->company_id); 
+        if(count($venta)==0 && count($ingreso)==0 && count($detalleinventario)==0){ 
+            if( $company->delete()){
         $path = public_path('logos/' . $company2->logo);
-            if (File::exists($path)) {   File::delete($path);   }
-
-        session()->flash('message','Proveedor o Cliente Eliminada');
-        $this->dispatchBrowserEvent('close-modal');
+            if (File::exists($path)) {   File::delete($path);   
+            }
+            session()->flash('message','Compañia Eliminada');
+            $this->dispatchBrowserEvent('close-modal');
+        }
+ 
+        }else{  
+            $company->status = 1;
+            $company->update();
+            session()->flash('message','Compañia Eliminada');
+            $this->dispatchBrowserEvent('close-modal');
+        }
+       
     }
 
     public function render()
     {
         
-        $companies = Company::orderBy('id','DESC')->paginate(10);
+        $companies = Company::orderBy('id','DESC')->where('status','=',0)->paginate(10);
         return view('livewire.admin.company.index',['companies' => $companies]);
     }
 }
